# swagger.yaml
swagger: '2.0'
info:
  description: ViruSafe Backend REST API
  version: 1.0.0
  title: ViruSafe Backend REST API
host: 'localhost:8080'
basePath: /
...
paths:
  /location/gps:
    post:
      tags:
        - location-controller
      summary: Submit GPS coordinates
      operationId: gpsLocationUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: JWT token
          required: true
          type: string
          x-example: Bearer eyJhbGciOiJIUzI1NiJ9.eyJjcmVhdGVkT24iOiIxNTY3NjUyNDAwIiwidXNlckd1aWQiOiJiYTBkNDdhYS0yYjYwLTQ5OTAtOGU5MS04MDIwYWU3ZTRkMTAiLCJqd3RTZWNyZXQiOiI3NjFiOGEyNi03N2YxLTQxNDEtOWVjNy0wZmZkMmEwZThkNjkiLCJwaG9uZU51bWJlciI6IjU1NTU1NTU1NTUiLCJleHAiOjE1OTkyNzQ4MDB9.DHgcQA0NAa7nySMD02dhhQKrcYP7LvC9AgJYLp-S0nQ
        - name: clientId
          in: header
          description: Client ID
          required: true
          type: string
          x-example: 3b3d3b02-e052-4289-827b-6c2f2e2a0430
        - in: body
          name: locationGpsDTO
          description: locationGpsDTO
          required: true
          schema:
            $ref: '#/definitions/LocationGpsDTO'
      responses:
        '201':
          description: Submitted successfully
        '401':
          description: Access unauthorized
          schema:
            $ref: '#/definitions/ErrorDTO'
        '403':
          description: Access forbidden
          schema:
            $ref: '#/definitions/ErrorDTO'
        '404':
          description: No corresponding user was found
          schema:
            $ref: '#/definitions/ErrorDTO'
        '412':
          description: Request argument validation has failed
          schema:
            $ref: '#/definitions/ValidationErrorDTO'
        '429':
          description: Request rate limit has been exceeded
          schema:
            $ref: '#/definitions/ErrorDTO'
        '500':
          description: An unexpected server error has occurred
          schema:
            $ref: '#/definitions/ErrorDTO'
      deprecated: false
  /location/proximity:
    post:
      tags:
        - location-controller
      summary: Submit devices in proximity
      operationId: proximityLocationUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: JWT token
          required: true
          type: string
          x-example: Bearer eyJhbGciOiJIUzI1NiJ9.eyJjcmVhdGVkT24iOiIxNTY3NjUyNDAwIiwidXNlckd1aWQiOiJiYTBkNDdhYS0yYjYwLTQ5OTAtOGU5MS04MDIwYWU3ZTRkMTAiLCJqd3RTZWNyZXQiOiI3NjFiOGEyNi03N2YxLTQxNDEtOWVjNy0wZmZkMmEwZThkNjkiLCJwaG9uZU51bWJlciI6IjU1NTU1NTU1NTUiLCJleHAiOjE1OTkyNzQ4MDB9.DHgcQA0NAa7nySMD02dhhQKrcYP7LvC9AgJYLp-S0nQ
        - name: clientId
          in: header
          description: Client ID
          required: true
          type: string
          x-example: 3b3d3b02-e052-4289-827b-6c2f2e2a0430
        - in: body
          name: locationProximityDTO
          description: locationProximityDTO
          required: true
          schema:
            $ref: '#/definitions/LocationProximityDTO'
      responses:
        '201':
          description: Submitted successfully
        '401':
          description: Access unauthorized
          schema:
            $ref: '#/definitions/ErrorDTO'
        '403':
          description: Access forbidden
          schema:
            $ref: '#/definitions/ErrorDTO'
        '404':
          description: No corresponding user was found
          schema:
            $ref: '#/definitions/ErrorDTO'
        '412':
          description: Request argument validation has failed
          schema:
            $ref: '#/definitions/ValidationErrorDTO'
        '429':
          description: Request rate limit has been exceeded
          schema:
            $ref: '#/definitions/ErrorDTO'
        '500':
          description: An unexpected server error has occurred
          schema:
            $ref: '#/definitions/ErrorDTO'
      deprecated: false
  /personalinfo:
    get:
      tags:
        - personal-information-controller
      summary: Get the current user's personal information
      operationId: getPersonalInformationUsingGET
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: JWT token
          required: true
          type: string
          x-example: Bearer eyJhbGciOiJIUzI1NiJ9.eyJjcmVhdGVkT24iOiIxNTY3NjUyNDAwIiwidXNlckd1aWQiOiJiYTBkNDdhYS0yYjYwLTQ5OTAtOGU5MS04MDIwYWU3ZTRkMTAiLCJqd3RTZWNyZXQiOiI3NjFiOGEyNi03N2YxLTQxNDEtOWVjNy0wZmZkMmEwZThkNjkiLCJwaG9uZU51bWJlciI6IjU1NTU1NTU1NTUiLCJleHAiOjE1OTkyNzQ4MDB9.DHgcQA0NAa7nySMD02dhhQKrcYP7LvC9AgJYLp-S0nQ
        - name: clientId
          in: header
          description: Client ID
          required: true
          type: string
          x-example: 3b3d3b02-e052-4289-827b-6c2f2e2a0430
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PersonalInformationResponseDTO'
        '401':
          description: Access unauthorized
          schema:
            $ref: '#/definitions/ErrorDTO'
        '403':
          description: Access forbidden
          schema:
            $ref: '#/definitions/ErrorDTO'
        '404':
          description: No corresponding user was found
          schema:
            $ref: '#/definitions/ErrorDTO'
        '500':
          description: An unexpected server error has occurred
          schema:
            $ref: '#/definitions/ErrorDTO'
      deprecated: false
    post:
      tags:
        - personal-information-controller
      summary: Submit user personal information
      operationId: addPersonalInformationUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: JWT token
          required: true
          type: string
          x-example: Bearer eyJhbGciOiJIUzI1NiJ9.eyJjcmVhdGVkT24iOiIxNTY3NjUyNDAwIiwidXNlckd1aWQiOiJiYTBkNDdhYS0yYjYwLTQ5OTAtOGU5MS04MDIwYWU3ZTRkMTAiLCJqd3RTZWNyZXQiOiI3NjFiOGEyNi03N2YxLTQxNDEtOWVjNy0wZmZkMmEwZThkNjkiLCJwaG9uZU51bWJlciI6IjU1NTU1NTU1NTUiLCJleHAiOjE1OTkyNzQ4MDB9.DHgcQA0NAa7nySMD02dhhQKrcYP7LvC9AgJYLp-S0nQ
        - name: clientId
          in: header
          description: Client ID
          required: true
          type: string
          x-example: 3b3d3b02-e052-4289-827b-6c2f2e2a0430
        - in: body
          name: personalInformationDTO
          description: personalInformationDTO
          required: true
          schema:
            $ref: '#/definitions/PersonalInformationRequestDTO'
      responses:
        '201':
          description: Submitted successfully
        '401':
          description: Access unauthorized
          schema:
            $ref: '#/definitions/ErrorDTO'
        '403':
          description: Access forbidden
          schema:
            $ref: '#/definitions/ErrorDTO'
        '404':
          description: No corresponding user was found
          schema:
            $ref: '#/definitions/ErrorDTO'
        '412':
          description: Request argument validation has failed
          schema:
            $ref: '#/definitions/ValidationErrorDTO'
        '429':
          description: Request rate limit has been exceeded
          schema:
            $ref: '#/definitions/ErrorDTO'
        '500':
          description: An unexpected server error has occurred
          schema:
            $ref: '#/definitions/ErrorDTO'
      deprecated: false
    delete:
      tags:
        - personal-information-controller
      summary: Delete the current user's personal information
      operationId: deletePersonalInformationUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: JWT token
          required: true
          type: string
          x-example: Bearer eyJhbGciOiJIUzI1NiJ9.eyJjcmVhdGVkT24iOiIxNTY3NjUyNDAwIiwidXNlckd1aWQiOiJiYTBkNDdhYS0yYjYwLTQ5OTAtOGU5MS04MDIwYWU3ZTRkMTAiLCJqd3RTZWNyZXQiOiI3NjFiOGEyNi03N2YxLTQxNDEtOWVjNy0wZmZkMmEwZThkNjkiLCJwaG9uZU51bWJlciI6IjU1NTU1NTU1NTUiLCJleHAiOjE1OTkyNzQ4MDB9.DHgcQA0NAa7nySMD02dhhQKrcYP7LvC9AgJYLp-S0nQ
        - name: clientId
          in: header
          description: Client ID
          required: true
          type: string
          x-example: 3b3d3b02-e052-4289-827b-6c2f2e2a0430
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Access unauthorized
          schema:
            $ref: '#/definitions/ErrorDTO'
        '403':
          description: Access forbidden
          schema:
            $ref: '#/definitions/ErrorDTO'
        '404':
          description: No corresponding user was found
          schema:
            $ref: '#/definitions/ErrorDTO'
        '429':
          description: Request rate limit has been exceeded
          schema:
            $ref: '#/definitions/ErrorDTO'
        '500':
          description: An unexpected server error has occurred
          schema:
            $ref: '#/definitions/ErrorDTO'
      deprecated: false
...
definitions:
...
  ErrorDTO:
    type: object
    properties:
      message:
        type: string
        example: Invalid request!
    title: ErrorDTO
  FieldValidationErrorDTO:
    type: object
    properties:
      fieldName:
        type: string
    title: FieldValidationErrorDTO
  Location:
    type: object
    properties:
      lat:
        type: number
        format: double
        example: 42.71
        minimum: -90
        maximum: 90
        exclusiveMinimum: false
        exclusiveMaximum: false
      lng:
        type: number
        format: double
        example: 23.32
        minimum: -180
        maximum: 180
        exclusiveMinimum: false
        exclusiveMaximum: false
    title: Location
  LocationGpsDTO:
    type: object
    required:
      - location
    properties:
      location:
        $ref: '#/definitions/Location'
      timestamp:
        type: integer
        format: int64
        example: 1567652400
    title: LocationGpsDTO
  LocationProximityDTO:
    type: object
    properties:
      location:
        $ref: '#/definitions/Location'
      proximities:
        type: array
        items:
          $ref: '#/definitions/ProximityDTO'
      timestamp:
        type: integer
        format: int64
        example: 1567652400
    title: LocationProximityDTO
  PersonalInformationRequestDTO:
    type: object
    properties:
      age:
        type: integer
        format: int32
        example: 33
        minimum: 14
        maximum: 150
        exclusiveMinimum: false
        exclusiveMaximum: false
      gender:
        type: string
        example: FEMALE
        enum:
          - MALE
          - FEMALE
      identificationNumber:
        type: string
        example: BG12345678
      identificationType:
        type: string
        example: PASSPORT
        enum:
          - EGN
          - LNCH
          - PASSPORT
      preExistingConditions:
        type: string
        example: Pre-existing conditions
        pattern: '^[a-zA-Z0-9 ,.()а-яА-Я-]{0,100}$'
    title: PersonalInformationRequestDTO
  PersonalInformationResponseDTO:
    type: object
    properties:
      age:
        type: integer
        format: int32
        example: 33
      gender:
        type: string
        example: FEMALE
        enum:
          - MALE
          - FEMALE
      identificationNumber:
        type: string
        example: BG12345678
      identificationType:
        type: string
        example: PASSPORT
        enum:
          - EGN
          - LNCH
          - PASSPORT
      phoneNumber:
        type: string
        example: 5555555555
      preExistingConditions:
        type: string
        example: Pre-existing conditions
    title: PersonalInformationResponseDTO
...
  ProximityDTO:
    type: object
    properties:
      distance:
        type: string
        example: 3.8
        pattern: '^[\d]+[.]{0,1}[\d]+$'
      uuid:
        type: string
        example: d70d5c29-30c0-415a-af6a-f7a2ccaa8c76
        pattern: '^[a-zA-Z0-9-]+$'
    title: ProximityDTO
...
  ValidationErrorDTO:
    type: object
    properties:
      message:
        type: string
        example: Request arguments validation failed!
      validationErrors:
        type: array
        items:
          $ref: '#/definitions/FieldValidationErrorDTO'
    title: ValidationErrorDTO